Refactor: AccountsView and separate AccountTradesView
